% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_sim_params}
\alias{add_sim_params}
\title{Add simulation parameters to results of powerms}
\usage{
add_sim_params(p)
}
\arguments{
\item{p}{A powerms_result object.}
}
\value{
The result as a dataframe with new columns corresponding to
the simulation factors (previously stored as an attributed).
}
\description{
Add simulation parameters to results of powerms
}
