% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_methods.R
\name{estimation_methods}
\alias{estimation_methods}
\alias{run_t_test}
\alias{run_mlm}
\title{Method of analysis}
\usage{
run_t_test(sdat, alpha = 0.05)

run_mlm(sdat, alpha = 0.05, psd_tau = 0.1, psd_sig_tau = 0.1, ncp = T)
}
\arguments{
\item{sdat}{site-aggregated dataset with site-level effect estimates \code{tau_j_hat} and site-level SEs \code{se_j}}

\item{alpha}{significance level for interval estimates}

\item{psd_tau}{sd for normal prior on average of site-level treatment effects}

\item{psd_sig_tau}{sd for half-normal prior on sd of site-level treatment effects}

\item{ncp}{use non-centered parameterization? Set to F if data are very highly informative}
}
\description{
These functions are the planned analysis functions used by the
power simulator.
}
