% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{moe_plot}
\alias{moe_plot}
\title{Plot results of powerms}
\usage{
moe_plot(p, x_axis, grouping = NULL)
}
\arguments{
\item{p}{Results from a powerms call}

\item{x_axis}{Unquoted name of simulation factor to plot on x-axis, e.g., \code{nbar}.
moe_plot() will marginalize over the remaining factors.}

\item{grouping}{Second factor to make a color, for multiple lines on the plot.}
}
\value{
ggplot object of desired plot
}
\description{
Given powerms() output, generate reasonable plot that shows average
margin of error, averaging over various simulation factors.
}
