% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{moe_plot_indiv}
\alias{moe_plot_indiv}
\title{plot results}
\usage{
moe_plot_indiv(p, sid = sid, grouping = NULL)
}
\arguments{
\item{p}{powermsresult object}

\item{sid}{unquoted site ID variable name}

\item{grouping}{(optional) unquoted grouping variable name}
}
\description{
plot results
}
