% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerms-package.R
\docType{package}
\name{powerms-package}
\alias{powerms-package}
\title{The 'powerms' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
