% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerms.R
\name{powerms}
\alias{powerms}
\title{conduct power analyses for a user-specified grid of settings}
\usage{
powerms(
  sim_data_method,
  se_method = "pooled",
  est_method,
  tx_var = "Z",
  outcome_var = "Y",
  site_id = "sid",
  num_sims = 100,
  parallel = F,
  ...
)
}
\arguments{
\item{se_method}{= "pooled" or "individual": how to compute
standard errors for the site-aggregated estimates}

\item{tx_var}{strings indicating variable names}

\item{outcome_var}{strings indicating variable names}

\item{sim_data_method:}{method for simulating data, default is the
sim_data() method in the package which uses fabricatr to simulate
a multisite trial}

\item{est_method:}{function for analyzing site-aggregated
estimates, built-in functions are run_mlm and run_t_test}

\item{site_id:}{strings indicating variable names, default "Z" "Y"
and "sid" (unfortunately curly-curly doesn't work with
parallelization so these are string variable names not just the
names)}

\item{num_sims:}{number of simulations to run}

\item{parallel:}{whether to turn on parallelization (via furrr)}

\item{...:}{additional arguments to sim_data_method. See
sim_data_method. Can input multiple arguments to a parameter,
e.g., n.bar = c(10, 20, 30), in which case the fully expanded
grid of parameters will be simulated and assembled}
}
\description{
NOTE: to use powerms locally while developing, need to
devtools::install() the full package! (furrr doesn't play nicely,
some globals issue...), as described in
https://github.com/DavisVaughan/furrr/issues/95
}
