% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerms_single.R
\name{powerms_single}
\alias{powerms_single}
\title{conduct power analysis for a single setting}
\usage{
powerms_single(
  sim_data_method,
  sim_data_args,
  se_method = "pooled",
  est_method,
  tx_var = "Z",
  outcome_var = "Y",
  site_id = "sid",
  num_sims = 100,
  parallel = F
)
}
\arguments{
\item{sim_data_args}{(unlisted)}
}
\value{
individual-level dataset
}
\description{
conduct power analysis for a single setting
}
