% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powermsresult.R
\name{powermsresult}
\alias{powermsresult}
\alias{sim_params}
\alias{is.powermsresult}
\alias{print.powermsresult}
\alias{head.powermsresult}
\alias{summary.powermsresult}
\title{Result object for results of powerms call}
\usage{
sim_params(x)

is.powermsresult(x)

\method{print}{powermsresult}(x, ...)

\method{head}{powermsresult}(x, n = 6L, ...)

\method{summary}{powermsresult}(object, ...)
}
\arguments{
\item{x}{object to take rows from}

\item{...}{extra options passed to print.powermsresult}

\item{n}{Number of rows to take}

\item{object}{object to summarize.}
}
\value{
The simulation parameters of a given simulation

is.powermsresult: TRUE if object is a powermsresult object.

print: No return value; prints results.

Data.frame of top rows

summary: No return value; prints results.
}
\description{
The powermsresult object is an S3 class that holds the results from \code{powerms()}.

It has several methods that pull different information from this object, and
some printing methods for getting nicely formatted results.

Calculate some summary statistics on the result object.

Calculate some summary statistics on the result object.
}
