% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate a dataset}
\usage{
sim_data(
  outcome = c("continuous", "binary"),
  intercept_dist = c("normal"),
  effect_dist = c("normal", "gamma"),
  J = 30,
  site_sizes = NULL,
  nbar = 10,
  vary_site_sizes = F,
  size_ratio = 1/3,
  site_ps = NULL,
  pbar = 0.5,
  vary_site_ps = F,
  alpha = 0,
  sig_alpha = sqrt(0.2),
  tau = 0.2,
  sig_tau = sqrt(0.2),
  rho = 0,
  a = NULL,
  b = NULL,
  cor_tau_n = 0,
  cor_tau_p = 0,
  ICC = 0.3
)
}
\description{
functions to simulate datasets:
}
\details{
\itemize{
\item input: all parameters of dataset
\item output: individual-level dataset
}
}
