% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sites.R
\name{summarize_sites}
\alias{summarize_sites}
\title{functions to produce table of site-level estimates (and standard errors),}
\usage{
summarize_sites(df, formula = NULL, se = c("pooled", "individual"))
}
\arguments{
\item{df}{individual-level dataset}

\item{formula}{formula indicating outcome, treatment, and site ID
column names from dataset, in the following form:
\verb{outcome ~ treatment | site ID}. Default \code{Y ~ Z | sid}.}

\item{se}{method for computing site-level standard errors,
options "pooled" or "individual"}
}
\value{
site-level dataset with estimated treatment effects \code{tau_hat},
standard errors \code{se}
}
\description{
functions to produce table of site-level estimates (and standard errors),
}
