% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_methods.R
\name{summarize_sites_fixed}
\alias{summarize_sites_fixed}
\title{same as summarize_sites, but uses:
\itemize{
\item site id sid
\item treatment Z
\item outcome Y
}}
\usage{
summarize_sites_fixed(df, se = c("pooled", "individual"))
}
\description{
same as summarize_sites, but uses:
\itemize{
\item site id sid
\item treatment Z
\item outcome Y
}
}
