% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerms_single.R
\name{summary_powerms_single}
\alias{summary_powerms_single}
\title{Make summary of output}
\usage{
summary_powerms_single(p)
}
\arguments{
\item{p}{output from powerms_single()}
}
\value{
prints useful results and description of powerms_single() run
}
\description{
Make summary of output
}
