% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{moe_plot}
\alias{moe_plot}
\title{Plot results of powerms}
\usage{
moe_plot(p, x_axis, grouping = NULL)
}
\arguments{
\item{p}{Results from a powerms call}

\item{x_axis}{Factor to make x-axis.  Name of variable in passed
dataframe or simulation parameter, e.g., \code{nbar}.  No string
quotations. Will margenalize over the rest of results.}

\item{grouping}{Second factor to make a color, for multiple lines on the plot.}
}
\value{
ggplot object of desired plot
}
\description{
given powerms() output, generate reasonable plot that shows average
margin of error, averaging over various simulation factors.
}
