% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerms.R
\name{powerms}
\alias{powerms}
\title{conduct power analyses for a user-specified grid of settings}
\usage{
powerms(
  sim_data_method,
  formula = NULL,
  se_method = "pooled",
  est_method,
  num_sims = 100,
  parallel = F,
  ...
)
}
\arguments{
\item{sim_data_method}{method for simulating data, default is the
sim_data() method in the package which uses fabricatr to simulate
a multisite trial}

\item{formula}{formula indicating outcome, treatment, and site ID
column names from simulated data, in the following form:
\verb{outcome ~ treatment | site ID}. Default \code{Y ~ Z | sid}.}

\item{se_method}{= "pooled" or "individual": how to compute
standard errors for the site-aggregated estimates}

\item{est_method}{function for analyzing site-aggregated estimates
\code{tau_hat} with standard errors \code{se}, outputs left (\code{ci_l}) and
right (\code{ci_r}) endpoints of interval estimates for each site.
Built-in functions are run_mlm() and run_t_test()}

\item{num_sims}{number of simulations to run}

\item{parallel}{whether to turn on parallelization (via furrr)}

\item{...}{additional arguments to sim_data_method. Can input
multiple arguments to a parameter, e.g., n.bar = c(10, 20, 30),
in which case the fully expanded grid of parameters will be
simulated and assembled}
}
\description{
NOTE: to use powerms locally while developing, need to
devtools::install() the full package! (furrr doesn't play nicely,
some globals issue...), as described in
https://github.com/DavisVaughan/furrr/issues/95
}
