% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_methods.R
\name{summarize_sites}
\alias{summarize_sites}
\title{functions to produce table of site-level estimates (and standard errors),}
\usage{
summarize_sites(
  df,
  tx_var,
  outcome_var,
  site_id,
  se = c("pooled", "individual")
)
}
\arguments{
\item{df}{an individual-level dataset.

@return  - input: individual-level dataset
Summarize sites

input df: needs sid, z, y
output df: sid, tau_hat, se

TODO: confirm that this works fine with binary outcomes}
}
\description{
functions to produce table of site-level estimates (and standard errors),
}
