% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_powerms_single}
\alias{summary_powerms_single}
\title{Make summary of output}
\usage{
summary_powerms_single(p)
}
\description{
input: powerms_single() output
output: useful results and description of powerms_single() run
}
